pyproject.toml
[tool.poetry]
name = "django-disguise"
version = "1.0.1"
description = ""
readme = "README.rst"
authors = ["Mikhail Porokhovnichenko <marazmiki@gmail.com>"]
license = "MIT"
repository = "https://github.com/marazmiki/django-disguise"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 1.11",
    "Framework :: Django :: 2.0",
    "Framework :: Django :: 2.1",
    "Framework :: Django :: 2.2",
    "Framework :: Django :: 3.0",
    "Framework :: Django :: 3.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
packages = [
    { include = "disguise" },
    { include = "disguise/**/*.py" },
]

[tool.poetry.dependencies]
python = "^3.5 || ^3.6 || ^3.7 || ^3.8 || ^3.9"

[tool.poetry.dev-dependencies]
pytest = "^6"
pytest-django = "^4"
pytest-isort = "^1"
pytest-flake8 = "^1.0.7"
pytest-env = "^0.6.2"
pytest-pythonpath = "^0.7.3"
pytest-cov = "^2.8.1"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
